Move `patch_and_save_config` into a separate non-commandline package so that importing it doesn't trigger CLI initialization.
