+++ splitgraph/commandline/engine.py
@@ -10,31 +10,11 @@ from tqdm import tqdm
 
 from splitgraph.__version__ import __version__
 from splitgraph.config import CONFIG, SG_CMD_ASCII, get_singleton
+from splitgraph.config.management import patch_and_save_config
 from splitgraph.exceptions import DockerUnavailableError, EngineSetupError
 from splitgraph.utils.docker import get_docker_client, copy_to_container
 
 DEFAULT_ENGINE = "default"
-
-
-def patch_and_save_config(config, patch):
-    from splitgraph.config.config import patch_config
-    from splitgraph.config.system_config import HOME_SUB_DIR
-    from splitgraph.config.export import overwrite_config
-    from pathlib import Path
-    import os
-
-    config_path = config["SG_CONFIG_FILE"]
-    if not config_path:
-        # Default to creating a config in the user's homedir rather than local.
-        config_dir = Path(os.environ["HOME"]) / Path(HOME_SUB_DIR)
-        config_path = config_dir / Path(".sgconfig")
-        logging.debug("No config file detected, creating one at %s" % config_path)
-        config_dir.mkdir(exist_ok=True, parents=True)
-    else:
-        logging.debug("Updating the existing config file at %s" % config_path)
-    new_config = patch_config(config, patch)
-    overwrite_config(new_config, config_path)
-    return str(config_path)
 
 
 def inject_config_into_engines(engine_prefix, config_path):
